#5/handle permissions

+++ src/Controller/TaskController.php
@@ -7,6 +7,7 @@ use App\Form\TaskType;
 use App\Handler\Forms\EntityFormHandler;
 use App\Repository\TaskRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -43,15 +44,41 @@ class TaskController extends AbstractController
 
     /**
      * @Route("/tasks", name="task_index")
+     * @param Request $request
      * @return Response
-
      */
-    public function index(): Response
+    public function index(Request $request): Response
     {
+        $options = [
+            "done" => $request->get('done'),
+        ];
+
         return $this->render(
             'task/index.html.twig', [
-                'tasks' => $this->taskRepository->findAll()
-            ]);
+            'tasks' => $this->taskRepository->findByUserQuery($this->getUser(), $options),
+            'showUsername' => false
+        ]);
+    }
+
+    /**
+     * @IsGranted("ROLE_ADMIN")
+     * @Route("/admin/tasks", name="admin_task_index")
+     * @param Request $request
+     * @return Response
+     */
+    public function adminIndex(Request $request): Response
+    {
+        $options = [
+            "done" => $request->get('done'),
+        ];
+        $anonymous = $request->get('anonymous');
+        $result = null !== $anonymous ? $this->taskRepository->findAnonymousQuery($options) : $this->taskRepository->findByQuery($options);
+
+        return $this->render(
+            'task/index.html.twig', [
+            'tasks' => $result,
+            'showUsername' => true,
+        ]);
     }
 
     /**
@@ -59,12 +86,12 @@ class TaskController extends AbstractController
      * @param Request $request
      * @return RedirectResponse|Response
      */
-    public function new(Request $request) : Response
+    public function new(Request $request): Response
     {
         $task = new Task();
         $task->setDone(false);
         $task->setUser($this->getUser());
-        if ($this->formHandler->handle($request, $task,TaskType::class)) {
+        if ($this->formHandler->handle($request, $task, TaskType::class)) {
             $this->addFlash('success', 'La tâche a été bien été ajoutée.');
             return $this->redirectToRoute('task_index');
         }
@@ -80,7 +107,8 @@ class TaskController extends AbstractController
      */
     public function edit(Task $task, Request $request)
     {
-        if ($this->formHandler->handle($request, $task,TaskType::class)) {
+        $this->denyAccessUnlessGranted('TASK_EDIT', $task, "Vous n'avez pas le droit d'éditer cette tâche");
+        if ($this->formHandler->handle($request, $task, TaskType::class)) {
             $this->addFlash('success', 'La tâche a été bien été modifiée.');
             return $this->redirectToRoute('task_index');
         }
@@ -98,6 +126,7 @@ class TaskController extends AbstractController
      */
     public function toggleTaskAction(Task $task)
     {
+        $this->denyAccessUnlessGranted('TASK_EDIT', $task, "Vous n'avez pas le droit d'éditer cette tâche");
         $task->setDone(!$task->isDone());
         $this->entityManager->flush();
 
@@ -113,6 +142,7 @@ class TaskController extends AbstractController
      */
     public function delete(Task $task)
     {
+        $this->denyAccessUnlessGranted('TASK_DELETE', $task, "Vous n'avez pas le droit de supprimer cette tâche");
         $this->entityManager->remove($task);
         $this->entityManager->flush();
